model changed done!

+++ Assignment3/rasterizer.cpp
@@ -428,7 +428,7 @@ int rst::rasterizer::get_index(int x, int y)
 void rst::rasterizer::set_pixel(const Vector2i &point, const Eigen::Vector3f &color)
 {
     //old index: auto ind = point.y() + point.x() * width;
-    int ind = (height-point.y())*width + point.x();
+    long ind = (height-point.y())*width + point.x();
     frame_buf[ind] = color;
 }
